+++ npm/webpack-dev-server/src/helpers/sourceRelativeWebpackModules.ts
@@ -66,8 +66,6 @@ export const cypressWebpackPath = (config: WebpackDevServerConfig) => {
 type FrameworkWebpackMapper = { [Property in Frameworks]: string | undefined }
 
 const frameworkWebpackMapper: FrameworkWebpackMapper = {
-  'create-react-app': 'react-scripts',
-  'vue-cli': '@vue/cli-service',
   react: undefined,
   vue: undefined,
   next: 'next',